requirements.txt
flask
pandas
numpy
matplotlib 
scikit-learn
scipy
tensorflow
keras
google-cloud-storage
flask-cors
holidays==0.24
prophet==1.1.1
plotly
statsmodels==0.13.3
